{23.06.10}-{1.0.7 version renewal}

+++ server/controller/mutation/mutation.note.js
@@ -5,10 +5,15 @@ import { Note } from "../../models/note.model.js";
 
 export const mutCreateNote = async (
   _,
-  { title, content, color, groupId },
+  { title, content, color, groupId, userId },
   { user }
 ) => {
   requireAuth(user);
+
+  if (user._id.toString() !== userId) {
+    throw new AuthenticationError("Unauthorized");
+  }
+
   const group = await Group.findById(groupId);
   if (!title || title.trim() === "") {
     throw new Error("Title is required");
@@ -21,7 +26,7 @@ export const mutCreateNote = async (
     title,
     content,
     color,
-    owner: user._id,
+    owner: userId,
     group: groupId,
   });
   return note.save();
@@ -29,14 +34,20 @@ export const mutCreateNote = async (
 
 export const mutUpdateNote = async (
   _,
-  { _id, title, content, color },
+  { _id, title, content, color, userId },
   { user }
 ) => {
   requireAuth(user);
+
+  if (user._id.toString() !== userId) {
+    throw new AuthenticationError("Unauthorized");
+  }
+
   const note = await Note.findById(_id);
   if (!note || note.owner.toString() !== user._id.toString()) {
     throw new AuthenticationError("Unauthorized");
   }
+
   if (!title || title.trim() === "") {
     throw new Error("Title is required");
   }
@@ -44,15 +55,22 @@ export const mutUpdateNote = async (
   return Note.findByIdAndUpdate(_id, { title, content, color }, { new: true });
 };
 
-export const mutDeleteNote = async (_, { _id }, { user }) => {
+export const mutDeleteNote = async (_, { _id, userId }, { user }) => {
   requireAuth(user);
+
+  if (user._id.toString() !== userId) {
+    throw new AuthenticationError("Unauthorized");
+  }
+
   const note = await Note.findById(_id);
   if (!note || note.owner.toString() !== user._id.toString()) {
     throw new AuthenticationError("Unauthorized");
   }
+
   const deletedNote = await Note.findByIdAndDelete(_id);
   if (!deletedNote) {
     throw new Error("Note not found or already deleted");
   }
+
   return deletedNote;
 };
